+++ src/lib/core/tests/TestCHIPErrorStr.cpp
@@ -235,7 +235,7 @@ static void CheckCoreErrorStr(nlTestSuite * inSuite, void * inContext)
     RegisterCHIPLayerErrorFormatter();
 
     // For each defined error...
-    for (int err : sContext)
+    for (long unsigned int err : sContext)
     {
         const char * errStr = ErrorStr(err);
         char expectedText[9];
