+++ src/system/tests/TestSystemErrorStr.cpp
@@ -68,7 +68,7 @@ static void CheckSystemErrorStr(nlTestSuite * inSuite, void * inContext)
     System::RegisterLayerErrorFormatter();
 
     // For each defined error...
-    for (int err : sContext)
+    for (long unsigned int err : sContext)
     {
         const char * errStr = ErrorStr(err);
         char expectedText[9];
